requirements.txt
google-genai
pyautogui
pyttsx3
speechrecognition
keyboard
pytesseract
pillow